refactor: extract session branching from AgentWorkspace
